# pyfunction: use METH_NOARGS for no-argument functions