@birkenfeld 2021-05-17
Funny - the no-args Python was already slower before in my benchmarks (from `tox -e py`).