@davidhewitt 2021-05-17
Curious! What OS / Python version are you measuring on?