@birkenfeld 2021-05-17
This is on Linux/Py3.9 and a early Ryzen 8-core.

`main` after merge:
```
---------------------------------------------------------------------------------------------- benchmark: 6 tests ---------------------------------------------------------------------------------------------
Name (time in ns)                Min                    Max                  Mean              StdDev                Median                IQR             Outliers  OPS (Kops/s)            Rounds  Iterations
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_no_args                101.6509 (1.0)         701.4653 (1.91)       112.4121 (1.0)       10.0457 (1.02)       111.8609 (1.0)       6.0468 (1.44)   15267;14966    8,895.8367 (1.0)      197981          43
test_no_args_py             125.1101 (1.23)        367.0299 (1.0)        137.9458 (1.23)       9.8136 (1.0)        137.0098 (1.22)      4.2099 (1.0)      9504;9103    7,249.2228 (0.81)      71013         100
test_args_and_kwargs_py     534.0000 (5.25)      4,429.3993 (12.07)      603.0717 (5.36)      39.9979 (4.08)       599.1023 (5.36)      4.8982 (1.16)     3616;6009    1,658.1777 (0.19)     144280          10
test_mixed_args_py          705.1014 (6.94)      4,477.4002 (12.20)      860.4933 (7.65)      60.2558 (6.14)       853.1002 (7.63)     24.9973 (5.94)     6949;7788    1,162.1241 (0.13)     116687          10
test_args_and_kwargs        809.9887 (7.97)     14,570.9782 (39.70)      933.1094 (8.30)     147.5474 (15.03)      929.9838 (8.31)     30.0060 (7.13)     1119;7158    1,071.6857 (0.12)     142026           1
test_mixed_args             950.0072 (9.35)     37,642.9816 (102.56)   1,094.5140 (9.74)     193.4248 (19.71)    1,089.9967 (9.74)     39.9887 (9.50)     1529;6015      913.6475 (0.10)     110486           1
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

before:
```
----------------------------------------------------------------------------------------------- benchmark: 6 tests -----------------------------------------------------------------------------------------------
Name (time in ns)                Min                     Max                  Mean                StdDev                Median                IQR             Outliers  OPS (Kops/s)            Rounds  Iterations
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_no_args                102.4384 (1.0)          961.8049 (1.03)       112.6693 (1.0)          8.7736 (1.20)       111.9511 (1.0)       5.3658 (3.16)   13426;12566    8,875.5310 (1.0)      196464          41
test_no_args_py             122.4098 (1.19)         937.1800 (1.0)        137.1851 (1.22)         7.2829 (1.0)        137.2100 (1.23)      1.7000 (1.0)      5494;7658    7,289.4191 (0.82)      58821         100
test_args_and_kwargs_py     535.9994 (5.23)      25,806.3024 (27.54)      609.1173 (5.41)       253.2263 (34.77)      595.9999 (5.32)     40.0003 (23.53)   1897;13039    1,641.7200 (0.18)     146822          10
test_mixed_args_py          698.9983 (6.82)       4,800.4003 (5.12)       768.7399 (6.82)        60.5227 (8.31)       763.9988 (6.82)     40.0993 (23.59)   10060;8183    1,300.8300 (0.15)     116266          10
test_args_and_kwargs        800.0061 (7.81)      23,622.0076 (25.21)      908.6107 (8.06)       185.8825 (25.52)      900.0069 (8.04)     79.9773 (47.05)    3022;6903    1,100.5813 (0.12)     133852           1
test_mixed_args             929.9838 (9.08)     510,565.0052 (544.79)   1,122.1239 (9.96)     2,061.4126 (283.05)   1,050.0080 (9.38)     51.0190 (30.01)    216;11263      891.1672 (0.10)     114417           1
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```